pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "rfwscopedaq"
authors = [
    {name = "Adam Carpenter", email = "adamc@jlab.org"},
]
description = "Software library for interacting with the RF Scope Waveform Dabase"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["rf", "RF", "waveform", "Waveform"]
license = {text = "JLab"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    'pyepics >= 3.5, < 4.0',
    'pandas >= 1.5, < 3.0',
    'numpy >= 1.24, < 3.0',
    'rfscopedb@git+https://github.com/JeffersonLab/rfscopedb#egg=v0.3.0'
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    'coverage >= 7.6, < 8.0',
    'pytest >= 8.3, < 9.0',
    'build >= 1.2, < 2.0'
]

[project.scripts]
RFWScopeDAQ = "rfwscopedaq.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["test/unit", "test/integration"]
norecursedirs = [".git", "node_modules", "venv"]

[tool.coverage.run]
branch = true
include = ['src/*']
command_line = '-m pytest'

# ... other project metadata fields as listed in:
#     https://packaging.python.org/en/latest/guides/writing-pyproject-toml/